backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pg8000==1.30.2
passlib==1.7.4
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
Pillow==10.0.1
PyJWT==2.8.0
reportlab==4.0.4
python-dateutil==2.8.2